+++ NEWS.md
@@ -17,6 +17,8 @@ to minimize the package dependencies of the production version.
 - Merged data now has an attribute 'synch_time' which will be used to update the synchronization information in the side bar.
 - user roles as defined in the config.yml will now be shown at several places in the application and will be documented alongside the user name if needed. 
 - Provided scaffolding for developers who plan to deploy on Posit Connect
+- Improved dataTable outputs
+- Move `shinymanager` to suggests since it is optional to use it.
 - Add functionality to mark which forms are required to review and which are not.
 - Improved required column verification of metadata columns when reading metadata using `get_metadata()`.
 
@@ -27,3 +29,4 @@ to minimize the package dependencies of the production version.
 - Fixed issue with the timeline figure when a subject had no completed events.
 - Fixed pre-processing bug where `get_metadata()` would error when a metadata.xlsx tab exist, but were empty
 - Fixed error when running `golem::run_dev()` when `clinsight` was not installed.
+- Fixed error that the app did not recognize the user roles when using `shinymanager`.